feat: move InternetBankingChooserItem class to PaymentMethodResources
